Face-Mask-Detection/requirements.txt
tensorflow
keras
imutils
numpy
opencv-python
matplotlib
scipy
pathlib
pymysql
pillow
scikit-learn